refactor and added tests

+++ app/database/interface.py
@@ -7,34 +7,63 @@ from app.database.database import SessionLocal
 db = SessionLocal()
 
 
-def create_conversation(event: dict):
-    db_user = db.query(models.User).filter(models.User.id==event['creator_id']).first()
-    if db_user is None:
-        db_user = models.User(id=event['creator_id'])
-        db.add(db_user)
-
-    db_conversation = db.query(models.Conversation).filter(models.Conversation.id == event['id']).first()
-    if db_conversation is None:
-        db_conversation = models.Conversation(id=event['id'], name=event['name'], created_at=datetime.datetime.fromisoformat(event['created_at']))
-        db_conversation.users.append(db_user)
-        db_conversation.owner = db_user
-        db.add(db_conversation)
-    db.commit()
-
+def event_contains_all_required_fields(event: dict, required_keys: list):
+    """
+    Verify that the event is complete
+    :param event:
+    :param required_keys:
+    :return:
+    """
+    for key in required_keys:
+        if key not in event:
+            return False
     return True
 
 
-def add_user_to_conversation(event: dict):
-    db_user = db.query(models.User).filter(models.User.id==event['user_id']).first()
-    db_participant = db.query(models.User).filter(models.User.id == event['participant_id']).first()
-    if db_participant is None:
-        db_participant = models.User(id=event['participant_id'])
-        db.add(db_participant)
+def create_conversation(event: dict):
+    """
+    Create a new conversation object from the event
+    :param event:
+    :return:
+    """
+    required_keys = ['id', 'name', 'creator_id', 'created_at']
+    if event_contains_all_required_fields(event, required_keys):
+        db_user = db.query(models.User).filter(models.User.id==event['creator_id']).first()
+        if db_user is None:
+            db_user = models.User(id=event['creator_id'])
+            db.add(db_user)
 
-    db_conversation = db.query(models.Conversation).filter(models.Conversation.id == event['conversation_id']).first()
-    if db_conversation is not None and db_conversation.owner == db_user:
-        db_conversation.users.append(db_participant)
+        db_conversation = db.query(models.Conversation).filter(models.Conversation.id == event['id']).first()
+        if db_conversation is None:
+            db_conversation = models.Conversation(
+                id=event['id'],
+                name=event['name'],
+                created_at=datetime.datetime.fromisoformat(event['created_at'])
+            )
+            db_conversation.users.append(db_user)
+            db_conversation.owner = db_user
+            db.add(db_conversation)
         db.commit()
+
+        return True
+    return False
+
+
+def add_user_to_conversation(event: dict):
+    required_keys = ['user_id', 'participant_id', 'conversation_id']
+    if event_contains_all_required_fields(event, required_keys):
+        db_user = db.query(models.User).filter(models.User.id==event['user_id']).first()
+        db_participant = db.query(models.User).filter(models.User.id == event['participant_id']).first()
+        if db_participant is None:
+            db_participant = models.User(id=event['participant_id'])
+            db.add(db_participant)
+
+        db_conversation = db.query(models.Conversation).filter(models.Conversation.id == event['conversation_id']).first()
+        if db_conversation is not None and db_conversation.owner == db_user and db_participant not in db_conversation.users:
+            db_conversation.users.append(db_participant)
+            db.commit()
+            return True
+    return False
 
 
 def add_message_to_conversation(event: dict):
@@ -60,7 +89,7 @@ def add_message_to_conversation(event: dict):
         db.commit()
 
 
-def remove_user_from_conversatin(event: dict):
+def remove_user_from_conversation(event: dict):
     db_user = db.query(models.User).filter(models.User.id==event['user_id']).first()
     db_conversation = db.query(models.Conversation).filter(models.Conversation.id==event['conversation_id']).first()
 
@@ -72,12 +101,16 @@ def remove_user_from_conversatin(event: dict):
 
 
 def delete_conversation(event: dict):
-    db_user = db.query(models.User).filter(models.User.id==event['user_id']).first()
-    db_conversation = db.query(models.Conversation).filter(models.Conversation.id==event['conversation_id']).first()
+    required_keys = ['user_id', 'conversation_id']
+    if event_contains_all_required_fields(event, required_keys):
+        db_user = db.query(models.User).filter(models.User.id==event['user_id']).first()
+        db_conversation = db.query(models.Conversation).filter(models.Conversation.id==event['conversation_id']).first()
 
-    if db_conversation.owner == db_user:
-        db.delete(db_conversation)
-        db.commit()
+        if db_conversation.owner == db_user:
+            db.delete(db_conversation)
+            db.commit()
+            return True
+    return False
